Move views to lib
