+++ lib/utils/date.spec.js
@@ -6,7 +6,7 @@ describe('#DateFormatter', () => {
       let formattedDate = dateUtil.formatDate('2018-08-14T12:01:00Z', 'YY')
       expect(formattedDate).toBe('18')
       formattedDate = dateUtil.formatDate('2018-08-14T12:01:00Z', 'MM-YY')
-      expect(formattedDate).toBe('Aug-18')      
+      expect(formattedDate).toBe('08-18')      
     })
   })
 })
