+++ lib/views/cycleTimeDataProvider.js
@@ -1,8 +1,8 @@
 const {TreeItem, TreeItemCollapsibleState} = require('vscode')
-const {model} = require('../../../model')
-const {commands} = require('../../commands')
+const {model} = require('../../model')
+const {commands} = require('../commands')
 const IterationPortalItem = require('./iterationPortalItem')
-const {testMatch, expressions} = require('../../utils/regularExpressions')
+const {testMatch, expressions} = require('../utils/regularExpressions')
 
 class CycleTimeTreeDataProvider {
   constructor(context) {
