+++ lib/views/cycleTimeDataProvider.spec.js
@@ -1,8 +1,8 @@
-jest.mock('../../../model')
+jest.mock('../../model')
 const cycleTimeDataProvider = require('./cycleTimeDataProvider')
-const {model} = require('../../../model')
+const {model} = require('../../model')
 const {TreeItem} = require('vscode')
-const {commands} = require('../../commands')
+const {commands} = require('../commands')
 
 const portal = {
   2018: {
