+++ lib/views/iterationPortalItem.js
@@ -1,4 +1,4 @@
-const {formatDate} = require('../../utils/date')
+const {formatDate} = require('../utils/date')
 
 
 class IterationPortalItem {
@@ -6,7 +6,7 @@ class IterationPortalItem {
     this.year = this._getYear(iteration.start)
     this.month = this._getMonth(iteration.start)
     this.iterationNumber = iteration.number
-    this.label = `${formatDate(iteration.start, 'DD-MM-YYYY')} to ${formatDate(iteration.finish, 'DD-MMM-YYYY')}`
+    this.label = `${formatDate(iteration.start, 'DD-MMM-YYYY')} to ${formatDate(iteration.finish, 'DD-MMM-YYYY')}`
   }
 
   _getYear(startDate) {
